delete moved pivot roller commands
