+++ src/main/java/com/stuypulse/robot/RobotContainer.java
@@ -11,13 +11,14 @@ import com.stuypulse.robot.commands.drive.DriveDefault;
 import com.stuypulse.robot.commands.drive.DriveJoystick;
 import com.stuypulse.robot.commands.leds.LEDApplyPattern;
 import com.stuypulse.robot.commands.leds.LEDDeafultCommand;
-import com.stuypulse.robot.commands.pivot.PivotAlgaeIntake;
-import com.stuypulse.robot.commands.pivot.PivotAlgaeOuttake;
-import com.stuypulse.robot.commands.pivot.PivotCoralOuttake;
 import com.stuypulse.robot.commands.pivot.PivotLower;
 import com.stuypulse.robot.commands.pivot.PivotRaise;
 import com.stuypulse.robot.commands.pivot.PivotStop;
-import com.stuypulse.robot.commands.pivot.PivotRollerStop;
+import com.stuypulse.robot.commands.pivot.roller.PivotAlgaeIntake;
+import com.stuypulse.robot.commands.pivot.roller.PivotAlgaeOuttake;
+import com.stuypulse.robot.commands.pivot.roller.PivotCoralOuttake;
+import com.stuypulse.robot.commands.pivot.roller.PivotRollerToDirection;
+import com.stuypulse.robot.commands.pivot.roller.PivotRollerStop;
 import com.stuypulse.robot.constants.Ports;
 import com.stuypulse.robot.constants.Settings;
 import com.stuypulse.robot.subsystems.climber.Climb;
