+++ src/main/java/com/stuypulse/robot/commands/auton/DoubleL1Auton.java
@@ -3,7 +3,7 @@ package com.stuypulse.robot.commands.auton;
 import com.stuypulse.robot.constants.Settings.LEDPatterns;
 import com.stuypulse.robot.commands.drive.DriveTank;
 import com.stuypulse.robot.commands.leds.LEDApplyPattern;
-import com.stuypulse.robot.commands.pivot.PivotCoralOuttake;
+import com.stuypulse.robot.commands.pivot.roller.PivotCoralOuttake;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitUntilCommand;
