--- src/main/java/com/stuypulse/robot/commands/pivot/PivotRollerStop.java
@@ -1,9 +1,0 @@
-package com.stuypulse.robot.commands.pivot;
-
-import com.stuypulse.robot.constants.Settings;
-
-public class PivotRollerStop extends PivotRollerToDirection {
-    public PivotRollerStop() {
-        super(Settings.Pivot.ROLLER_STOP_SPEED.getAsDouble());
-    }
-}
